reorganized styling, updates to photo page
