+++ package.json
@@ -11,6 +11,7 @@
     "react-dom": "^17.0.2",
     "react-router-dom": "^6.3.0",
     "react-scripts": "5.0.0",
+    "sass": "^1.51.0",
     "typescript": "^4.6.2"
   },
   "scripts": {
