+++ src/_app.scss
@@ -38,29 +38,22 @@ h2 {
   margin-bottom: 0.125rem;
 }
 
-p,
-a {
-  margin: 0;
-}
 a {
   color: #939393;
   text-decoration: none;
 }
+
 h1:hover,
 a:hover {
   color: #f0152f;
 }
+
 ul {
   padding: 0;
   margin: 0;
 }
+
 li {
   margin: 0 0 0.5rem 0;
   list-style: none;
-}
-
-.photoCard img {
-  width: 100%;
-  max-height: 20rem;
-  object-fit: cover;
 }
