+++ src/components/_index.scss
@@ -1,0 +1,1 @@
+@use "PhotoCard/PhotoCard";
